requirements.txt
pyaudio
pydirectinput~=1.0.4
keyboard
PySynonym~=0.0.9
pypiwin32~=223
pyttsx3
pyglet~=1.5.5
winwifi~=1.0.3
PySimpleGUI~=4.19.0
pyautogui~=0.9.50
matplotlib~=3.2.1
Pillow~=8.2.0
pyperclip~=1.8.2
bs4~=0.0.1
beautifulsoup4~=4.9.3
pywin32~=301
numpy~=1.20.3